+++ app/todos/page.tsx
@@ -7,6 +7,7 @@ import { getCsrfToken } from '@/lib/csrf';
 import { useUser } from '@/lib/auth';
 import { logout } from '@/lib/auth';
 import LogoutButton from './components/LogoutButton';
+import TaskForm from './components/TaskForm';
 import axios from 'axios';
 import dayjs from 'dayjs';
 import utc from 'dayjs/plugin/utc';
@@ -25,10 +26,7 @@ export default function TodosPage() {
   const router = useRouter();
   const { user, isLoading, isError } = useUser();
   const [tasks, setTasks] = useState<Task[]>([]);
-  const [newTask, setNewTask] = useState('');
-  const [dueDate, setDueDate] = useState<string>('');
 
-  const [isPosting, setIsPosting] = useState(false);
   const [isToggling, setIsToggling] = useState(false);
   const [isDeleting, setIsDeleting] = useState(false);
   const [loggingOut, setLoggingOut] = useState(false);
@@ -54,36 +52,6 @@ export default function TodosPage() {
   const fetchTasks = async () => {
     const res = await api.get('/api/tasks');
     setTasks(res.data);
-  };
-
-  const addTask = async (e: React.FormEvent) => {
-    e.preventDefault();
-    if (isPosting || !newTask.trim()) return;
-    if (dueDate && dayjs(dueDate).isBefore(dayjs())) {
-      alert('期限は現在より未来の日時を指定してください。');
-      return;
-    }
-    setIsPosting(true);
-    try {
-      await getCsrfToken();
-      await api.post('/api/tasks', {
-        title: newTask,
-        due_date: dueDate ? dayjs(dueDate).tz('Asia/Tokyo').format('YYYY-MM-DDTHH:mm:ssZ') : null,
-      });
-      setNewTask('');
-      setDueDate('');
-      fetchTasks();
-    } catch (error: unknown) {
-      if (axios.isAxiosError(error) && error.response && error.response.status === 422) {
-        const errors = error.response.data.errors as Record<string, string[]>;
-        const messages = Object.values(errors).flat();
-        setErrorMessages(messages);
-      } else {
-        console.error('Unexpected error:', error);
-      }
-    } finally {
-      setIsPosting(false);
-    }
   };
 
   const toggleTask = async (task: Task) => {
@@ -174,24 +142,7 @@ export default function TodosPage() {
         <LogoutButton onLogout={handleLogout} loggingOut={loggingOut} />
       </div>
 
-      <form onSubmit={addTask} className="flex mb-4 space-x-2">
-        <input
-          className="flex-grow border px-3 py-2"
-          placeholder="新しいタスクを追加"
-          value={newTask}
-          onChange={(e) => setNewTask(e.target.value)}
-        />
-        <input
-          type="datetime-local"
-          className="flex-grow px-3 py-2"
-          value={dueDate}
-          onChange={(e) => setDueDate(e.target.value)}
-          min={dayjs().format('YYYY-MM-DDTHH:mm')}  // 過去禁止
-        />
-        <button type="submit" className="bg-blue-500 text-white px-4 py-2" disabled={isPosting}>
-          追加
-        </button>
-      </form>
+      <TaskForm onAdd={fetchTasks} setErrorMessages={setErrorMessages} />
       {errorMessages.length > 0 && (
         <div className="text-red-500 mb-4">
           {errorMessages.map((msg, index) => (
